Docks in progress

+++ src/course/course.controller.ts
@@ -11,11 +11,13 @@ import {
   BadRequestException,
 } from '@nestjs/common';
 import { CourseService } from './course.service';
-import { Prisma } from '@prisma/client';
+import { CreateCourseDto } from './dto/create-course.dto';
+import { UpdateCourseDto } from './dto/update-course.dto';
 import { FileInterceptor } from '@nestjs/platform-express';
 import * as fs from 'fs';
 import * as path from 'path';
 import { resizeAndOptimizeImage } from 'libs/common/src/helpers/resize-image-helper';
+
 @Controller('courses')
 export class CourseController {
   constructor(private readonly courseService: CourseService) {}
@@ -23,11 +25,12 @@ export class CourseController {
   @Post()
   @UseInterceptors(FileInterceptor('image'))
   async create(
-    @Body() createCourseDto: Prisma.CourseCreateInput,
+    @Body() createCourseDto: CreateCourseDto,
     @UploadedFile() image: Express.Multer.File,
   ) {
     let newCourse = createCourseDto;
 
+    // Handle image if provided
     if (image) {
       const imageDir = path.join(__dirname, '..', '..', '..', 'images');
       if (!fs.existsSync(imageDir)) {
@@ -58,24 +61,26 @@ export class CourseController {
   @UseInterceptors(FileInterceptor('image'))
   async update(
     @Param('id') id: string,
-    @Body() updateCourseDto: Prisma.CourseUpdateInput,
+    @Body() updateCourseDto: UpdateCourseDto,
     @UploadedFile() image: Express.Multer.File,
   ) {
     let updatedCourse = updateCourseDto;
-    const currnetCourse = await this.courseService.findOne(id);
-    if (!currnetCourse) {
+    const currentCourse = await this.courseService.findOne(id);
+    if (!currentCourse) {
       throw new BadRequestException('Course not found');
     }
 
+    // Handle image update if provided
     if (image) {
       const imageDir = path.join(__dirname, '..', '..', '..', 'images');
       try {
-        //Delete old image
-        const oldImagePath = path.join(imageDir, currnetCourse.courseImageUrl);
+        // Delete old image if it exists
+        const oldImagePath = path.join(imageDir, currentCourse.courseImageUrl);
         if (fs.existsSync(oldImagePath)) {
           fs.unlinkSync(oldImagePath);
         }
-        //Create new
+
+        // Create and save new image
         const optimizedImageBuffer = await resizeAndOptimizeImage(image.buffer);
         const fileExt = path.extname(image.originalname);
         const courseImageName = `courseImage_${Date.now()}${fileExt}`;
@@ -84,7 +89,7 @@ export class CourseController {
         fs.writeFileSync(courseImagePath, optimizedImageBuffer);
         updatedCourse.courseImageUrl = courseImageName;
       } catch (error) {
-        console.error('Error file:', error);
+        console.error('Error updating image:', error);
       }
     }
 
@@ -92,7 +97,19 @@ export class CourseController {
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  async remove(@Param('id') id: string) {
+    const currentCourse = await this.courseService.findOne(id);
+    if (!currentCourse) {
+      throw new BadRequestException('Course not found');
+    }
+
+    // Delete image associated with the course if it exists
+    const imageDir = path.join(__dirname, '..', '..', '..', 'images');
+    const courseImagePath = path.join(imageDir, currentCourse.courseImageUrl);
+    if (fs.existsSync(courseImagePath)) {
+      fs.unlinkSync(courseImagePath);
+    }
+
     return this.courseService.remove(id);
   }
 }
